requirements.txt
streamlit
streamlit_extras
pathlib
pillow
torch
opencv-python
openai-clip
sqlite3